Extract SectionExecutionFactory from SqlQueryScheduler
