+++ presto-main/src/main/java/com/facebook/presto/execution/SqlQueryExecution.java
@@ -23,14 +23,12 @@ import com.facebook.presto.execution.StateMachine.StateChangeListener;
 import com.facebook.presto.execution.buffer.OutputBuffers;
 import com.facebook.presto.execution.buffer.OutputBuffers.OutputBufferId;
 import com.facebook.presto.execution.scheduler.ExecutionPolicy;
-import com.facebook.presto.execution.scheduler.NodeScheduler;
+import com.facebook.presto.execution.scheduler.SectionExecutionFactory;
 import com.facebook.presto.execution.scheduler.SplitSchedulerStats;
 import com.facebook.presto.execution.scheduler.SqlQueryScheduler;
 import com.facebook.presto.execution.warnings.WarningCollector;
-import com.facebook.presto.failureDetector.FailureDetector;
 import com.facebook.presto.memory.VersionedMemoryPoolId;
 import com.facebook.presto.metadata.Metadata;
-import com.facebook.presto.operator.ForScheduler;
 import com.facebook.presto.security.AccessControl;
 import com.facebook.presto.server.BasicQueryInfo;
 import com.facebook.presto.spi.ConnectorId;
@@ -49,7 +47,6 @@ import com.facebook.presto.sql.analyzer.QueryExplainer;
 import com.facebook.presto.sql.parser.SqlParser;
 import com.facebook.presto.sql.planner.InputExtractor;
 import com.facebook.presto.sql.planner.LogicalPlanner;
-import com.facebook.presto.sql.planner.NodePartitioningManager;
 import com.facebook.presto.sql.planner.OutputExtractor;
 import com.facebook.presto.sql.planner.PartitioningHandle;
 import com.facebook.presto.sql.planner.Plan;
@@ -76,7 +73,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
 
@@ -105,20 +101,15 @@ public class SqlQueryExecution
     private final Metadata metadata;
     private final SqlParser sqlParser;
     private final SplitManager splitManager;
-    private final NodePartitioningManager nodePartitioningManager;
-    private final NodeScheduler nodeScheduler;
     private final List<PlanOptimizer> planOptimizers;
     private final PlanFragmenter planFragmenter;
     private final RemoteTaskFactory remoteTaskFactory;
     private final LocationFactory locationFactory;
-    private final int scheduleSplitBatchSize;
     private final ExecutorService queryExecutor;
-    private final ScheduledExecutorService schedulerExecutor;
-    private final FailureDetector failureDetector;
+    private final SectionExecutionFactory sectionExecutionFactory;
 
     private final AtomicReference<SqlQueryScheduler> queryScheduler = new AtomicReference<>();
     private final AtomicReference<Plan> queryPlan = new AtomicReference<>();
-    private final NodeTaskMap nodeTaskMap;
     private final ExecutionPolicy executionPolicy;
     private final SplitSchedulerStats schedulerStats;
     private final Analysis analysis;
@@ -138,17 +129,12 @@ public class SqlQueryExecution
             AccessControl accessControl,
             SqlParser sqlParser,
             SplitManager splitManager,
-            NodePartitioningManager nodePartitioningManager,
-            NodeScheduler nodeScheduler,
             List<PlanOptimizer> planOptimizers,
             PlanFragmenter planFragmenter,
             RemoteTaskFactory remoteTaskFactory,
             LocationFactory locationFactory,
-            int scheduleSplitBatchSize,
             ExecutorService queryExecutor,
-            ScheduledExecutorService schedulerExecutor,
-            FailureDetector failureDetector,
-            NodeTaskMap nodeTaskMap,
+            SectionExecutionFactory sectionExecutionFactory,
             QueryExplainer queryExplainer,
             ExecutionPolicy executionPolicy,
             SplitSchedulerStats schedulerStats,
@@ -161,22 +147,15 @@ public class SqlQueryExecution
             this.metadata = requireNonNull(metadata, "metadata is null");
             this.sqlParser = requireNonNull(sqlParser, "sqlParser is null");
             this.splitManager = requireNonNull(splitManager, "splitManager is null");
-            this.nodePartitioningManager = requireNonNull(nodePartitioningManager, "nodePartitioningManager is null");
-            this.nodeScheduler = requireNonNull(nodeScheduler, "nodeScheduler is null");
             this.planOptimizers = requireNonNull(planOptimizers, "planOptimizers is null");
             this.planFragmenter = requireNonNull(planFragmenter, "planFragmenter is null");
             this.locationFactory = requireNonNull(locationFactory, "locationFactory is null");
             this.queryExecutor = requireNonNull(queryExecutor, "queryExecutor is null");
-            this.schedulerExecutor = requireNonNull(schedulerExecutor, "schedulerExecutor is null");
-            this.failureDetector = requireNonNull(failureDetector, "failureDetector is null");
-            this.nodeTaskMap = requireNonNull(nodeTaskMap, "nodeTaskMap is null");
+            this.sectionExecutionFactory = requireNonNull(sectionExecutionFactory, "sectionExecutionFactory is null");
             this.executionPolicy = requireNonNull(executionPolicy, "executionPolicy is null");
             this.schedulerStats = requireNonNull(schedulerStats, "schedulerStats is null");
             this.statsCalculator = requireNonNull(statsCalculator, "statsCalculator is null");
             this.costCalculator = requireNonNull(costCalculator, "costCalculator is null");
-
-            checkArgument(scheduleSplitBatchSize > 0, "scheduleSplitBatchSize must be greater than 0");
-            this.scheduleSplitBatchSize = scheduleSplitBatchSize;
 
             requireNonNull(query, "query is null");
             requireNonNull(session, "session is null");
@@ -497,21 +476,15 @@ public class SqlQueryExecution
                 stateMachine,
                 locationFactory,
                 outputStagePlan,
-                nodePartitioningManager,
-                nodeScheduler,
                 remoteTaskFactory,
                 splitSourceFactory,
                 stateMachine.getSession(),
                 plan.isSummarizeTaskInfos(),
-                scheduleSplitBatchSize,
                 queryExecutor,
-                schedulerExecutor,
-                failureDetector,
                 rootOutputBuffers,
-                nodeTaskMap,
                 executionPolicy,
                 schedulerStats,
-                metadata);
+                sectionExecutionFactory);
 
         queryScheduler.set(scheduler);
 
@@ -660,13 +633,10 @@ public class SqlQueryExecution
             implements QueryExecutionFactory<QueryExecution>
     {
         private final SplitSchedulerStats schedulerStats;
-        private final int scheduleSplitBatchSize;
         private final Metadata metadata;
         private final AccessControl accessControl;
         private final SqlParser sqlParser;
         private final SplitManager splitManager;
-        private final NodePartitioningManager nodePartitioningManager;
-        private final NodeScheduler nodeScheduler;
         private final List<PlanOptimizer> planOptimizers;
         private final PlanFragmenter planFragmenter;
         private final RemoteTaskFactory remoteTaskFactory;
@@ -674,9 +644,7 @@ public class SqlQueryExecution
         private final QueryExplainer queryExplainer;
         private final LocationFactory locationFactory;
         private final ExecutorService queryExecutor;
-        private final ScheduledExecutorService schedulerExecutor;
-        private final FailureDetector failureDetector;
-        private final NodeTaskMap nodeTaskMap;
+        private final SectionExecutionFactory sectionExecutionFactory;
         private final Map<String, ExecutionPolicy> executionPolicies;
         private final ClusterSizeMonitor clusterSizeMonitor;
         private final StatsCalculator statsCalculator;
@@ -689,16 +657,12 @@ public class SqlQueryExecution
                 SqlParser sqlParser,
                 LocationFactory locationFactory,
                 SplitManager splitManager,
-                NodePartitioningManager nodePartitioningManager,
-                NodeScheduler nodeScheduler,
                 PlanOptimizers planOptimizers,
                 PlanFragmenter planFragmenter,
                 RemoteTaskFactory remoteTaskFactory,
                 TransactionManager transactionManager,
                 @ForQueryExecution ExecutorService queryExecutor,
-                @ForScheduler ScheduledExecutorService schedulerExecutor,
-                FailureDetector failureDetector,
-                NodeTaskMap nodeTaskMap,
+                SectionExecutionFactory sectionExecutionFactory,
                 QueryExplainer queryExplainer,
                 Map<String, ExecutionPolicy> executionPolicies,
                 SplitSchedulerStats schedulerStats,
@@ -708,22 +672,17 @@ public class SqlQueryExecution
         {
             requireNonNull(config, "config is null");
             this.schedulerStats = requireNonNull(schedulerStats, "schedulerStats is null");
-            this.scheduleSplitBatchSize = config.getScheduleSplitBatchSize();
             this.metadata = requireNonNull(metadata, "metadata is null");
             this.accessControl = requireNonNull(accessControl, "accessControl is null");
             this.sqlParser = requireNonNull(sqlParser, "sqlParser is null");
             this.locationFactory = requireNonNull(locationFactory, "locationFactory is null");
             this.splitManager = requireNonNull(splitManager, "splitManager is null");
-            this.nodePartitioningManager = requireNonNull(nodePartitioningManager, "nodePartitioningManager is null");
-            this.nodeScheduler = requireNonNull(nodeScheduler, "nodeScheduler is null");
             requireNonNull(planOptimizers, "planOptimizers is null");
             this.planFragmenter = requireNonNull(planFragmenter, "planFragmenter is null");
             this.remoteTaskFactory = requireNonNull(remoteTaskFactory, "remoteTaskFactory is null");
             this.transactionManager = requireNonNull(transactionManager, "transactionManager is null");
             this.queryExecutor = requireNonNull(queryExecutor, "queryExecutor is null");
-            this.schedulerExecutor = requireNonNull(schedulerExecutor, "schedulerExecutor is null");
-            this.failureDetector = requireNonNull(failureDetector, "failureDetector is null");
-            this.nodeTaskMap = requireNonNull(nodeTaskMap, "nodeTaskMap is null");
+            this.sectionExecutionFactory = requireNonNull(sectionExecutionFactory, "sectionExecutionFactory is null");
             this.queryExplainer = requireNonNull(queryExplainer, "queryExplainer is null");
             this.executionPolicies = requireNonNull(executionPolicies, "schedulerPolicies is null");
             this.clusterSizeMonitor = requireNonNull(clusterSizeMonitor, "clusterSizeMonitor is null");
@@ -758,17 +717,12 @@ public class SqlQueryExecution
                     accessControl,
                     sqlParser,
                     splitManager,
-                    nodePartitioningManager,
-                    nodeScheduler,
                     planOptimizers,
                     planFragmenter,
                     remoteTaskFactory,
                     locationFactory,
-                    scheduleSplitBatchSize,
                     queryExecutor,
-                    schedulerExecutor,
-                    failureDetector,
-                    nodeTaskMap,
+                    sectionExecutionFactory,
                     queryExplainer,
                     executionPolicy,
                     schedulerStats,
